enable react 19 SSR polyfill
