+++ package.json
@@ -26,7 +26,7 @@
     "@eslint/js": "9.15.0",
     "@tanstack/config": "^0.7.0",
     "@types/node": "^22.0.0",
-    "@types/react": "^18.3.13",
+    "@types/react": "^19.0.2",
     "@vitejs/plugin-react": "^4.2.1",
     "eslint": "^9.15.0",
     "eslint-import-resolver-typescript": "^3.6.1",
