+++ packages/tuono/src/index.ts
@@ -1,14 +1,12 @@
-export { Helmet as Head } from 'react-helmet-async'
-
 export {
-  createRoute,
-  createRootRoute,
-  createRouter,
-  Link,
-  RouterProvider,
-  dynamic,
-  useRouter,
-  __tuono__internal__lazyLoadComponent,
+	createRoute,
+	createRootRoute,
+	createRouter,
+	Link,
+	RouterProvider,
+	dynamic,
+	useRouter,
+	__tuono__internal__lazyLoadComponent,
 } from 'tuono-router'
 
 export type { TuonoProps } from './types'
