--- packages/tuono/src/react-meta-tags.d.ts
@@ -1,2 +1,0 @@
-declare module 'react-meta-tags'
-declare module 'react-meta-tags/server'
