+++ packages/tuono/src/ssr/polyfills/messageEvent.ts
@@ -1,0 +1,12 @@
+function Event(type) {
+	this.type = type
+	this.target = undefined
+}
+
+export function MessageEventPolyfill(type, options) {
+	Event.call(this, type)
+	this.data = options.data
+	this.lastEventId = options.lastEventId
+}
+
+MessageEvent.prototype = Object.create(Event.prototype)
